logout with ajax call

+++ src/controllers/Item.php
@@ -14,11 +14,11 @@ class Item
 
     public function create()
     {
-        if (!isLoggedIn() || !isSeller()) {
-            header('location: ' . URLROOT . '/', true, 303);
-            die(UNAUTHORIZED_ACCESS);
-        }
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+            if (!isLoggedIn() || !isSeller()) {
+                header('location: ' . URLROOT . '/', true, 303);
+                die(UNAUTHORIZED_ACCESS);
+            }
             $itemRequest = [
                 'brand' => $_POST['brand'],
                 'model' => $_POST['model'],
@@ -34,17 +34,23 @@ class Item
                 header('location: ' . URLROOT . '/admin', true, 303);
             else
                 die(SOMETHING_WENT_WRONG);
+        } else {
+            die(UNAUTHORIZED_ACCESS);
         }
     }
 
     public function delete($params)
     {
-        if (!isLoggedIn() || !isAdmin()) {
-            header('location: ' . URLROOT . '/', true, 303);
+        if ($_SERVER['REQUEST_METHOD'] === 'POST') {
+            if (!isLoggedIn() || !isAdmin()) {
+                header('location: ' . URLROOT . '/', true, 303);
+                die(UNAUTHORIZED_ACCESS);
+            }
+            $id = $params['id'];
+            $this->size->deleteItem($id);
+            header('location: ' . URLROOT . '/admin', true, 303);
+        } else {
             die(UNAUTHORIZED_ACCESS);
         }
-        $id = $params['id'];
-        $this->size->deleteItem($id);
-        header('location: ' . URLROOT . '/admin', true, 303);
     }
 }
